Final change to ImageClass. New class made -> DFTFilter within ImageClass ( as suggested by kscottz ) Supports reuse of same filter for multiple images!
